requirements-dev.txt
# Build tools
setuptools
setuptools_rust

# Lint tools
flake8 
pep440

# Test tools
matplotlib 
pytest
pytest-xdist

# qiskit-aer
qiskit-aer
